fix issue when jq transform has a filter, cleanup tests
